+++ arm9/source/romBrowser/viewModels/RomBrowserViewModel.cpp
@@ -30,6 +30,7 @@ RomBrowserViewModel::RomBrowserViewModel(IRomBrowserController* romBrowserContro
         }
     }
     filterSortParams.favoritesOnly = romBrowserController->IsFavoritesFilterEnabled();
+    filterSortParams.completedOnly = romBrowserController->IsCompletedFilterEnabled();
     filterSortParams.gameDataService = romBrowserController->GetGameDataService();
     u64 startTick = gTickCounter.GetValue();
     const auto& sdFolder = romBrowserController->GetSdFolder();