+++ arm9/source/services/gamedata/JsonGameDataService.h
@@ -9,6 +9,7 @@ public:
 
     const GameDataEntry* GetEntry(const char* fileName, const char* gameCode = nullptr) const override;
     void ToggleFavorite(const char* fileName, const char* gameCode = nullptr) override;
+    void ToggleCompleted(const char* fileName, const char* gameCode = nullptr) override;
     void RecordLaunch(const char* fileName, const char* gameCode,
         const char* fullPath, const char* lastPlayedDateTime) override;
     void RemoveEntry(const char* fileName, const char* gameCode = nullptr) override;