+++ arm9/source/settings/viewModels/ThemeListItemViewModel.h
@@ -18,9 +18,16 @@ public:
 
     void ToggleFavorite() override { }
 
+    void ToggleCompleted() override { }
+
     void SetIndex(int index) override
     {
         _index = index;
+    }
+
+    int GetIndex() const override
+    {
+        return _index;
     }
 
     void SetQueueTask(QueueTask<void> queueTask) override